+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `dxm monitor install` to install a third-party monitor.
 - `dxm monitor remove` to remove the third-party monitor if exists.
 - `dxm update --monitor` to update the third-party monitor if exists.
+- `dxm add --git` to install third-party resources through any Git repository.
 
 ### Changed