+++ crates/dxm-resources/CHANGELOG.md
@@ -13,10 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `VacatedDir` to move a directory to a different location to be later brought
   back or deleted.
+- `format_git_url` to create a download URL for Git resources.
 
 ### Changed
 
 - Updated dependencies.
+- Changed `install` to return an updated download URL if the given source URL is
+  missing details.
+- Updated `resolve` to resolve and return `DownloadSource::Git` for URLs
+  prefixed with `git+[rev]+`.
 
 ### Removed